Migrate toggle from /docs/components/ to /components/
